Club - An Cross Chain Interactive Video Protocol and Network
We are applying this grant to enable real time messaging and video chat available to both Polkadot and Ethereum users based on Club. 
Club is an interactive real time video protocol and network which can be used in most blockchain today to provide real time video communication for the wallet users and blockchain developers.